+++ backend/src/main/java/ca/metricalsky/winston/service/pull/estimate/ChannelQuotaCostEstimator.java
@@ -1,0 +1,14 @@
+package ca.metricalsky.winston.service.pull.estimate;
+
+import ca.metricalsky.winston.api.model.PullChannelOperation;
+import org.springframework.stereotype.Service;
+
+@Service
+public class ChannelQuotaCostEstimator implements QuotaCostEstimator<PullChannelOperation> {
+
+    @Override
+    public int estimateQuotaCost(PullChannelOperation operation) {
+        operation.setEstimatedCost(1);
+        return operation.getEstimatedCost();
+    }
+}