+++ backend/src/main/java/ca/metricalsky/winston/service/pull/estimate/QuotaCostEstimator.java
@@ -1,0 +1,6 @@
+package ca.metricalsky.winston.service.pull.estimate;
+
+public interface QuotaCostEstimator<T> {
+
+    int estimateQuotaCost(T operation);
+}