+++ backend/src/test/java/ca/metricalsky/winston/service/fetch/FetchServiceTest.java
@@ -1,6 +1,6 @@
 package ca.metricalsky.winston.service.fetch;
 
-import ca.metricalsky.winston.api.model.FetchRequest;
+import ca.metricalsky.winston.config.properties.youtube.YouTubeConfig;
 import ca.metricalsky.winston.entity.fetch.FetchOperationEntity;
 import ca.metricalsky.winston.entity.fetch.FetchOperationEntity.Type;
 import ca.metricalsky.winston.entity.fetch.FetchRequestEntity;
@@ -8,8 +8,6 @@ import ca.metricalsky.winston.events.PublisherException;
 import ca.metricalsky.winston.events.SsePublisher;
 import ca.metricalsky.winston.events.SsePublisherHolder;
 import ca.metricalsky.winston.exception.AppException;
-import ca.metricalsky.winston.mapper.entity.FetchRequestMapper;
-import ca.metricalsky.winston.repository.fetch.FetchRequestRepository;
 import ca.metricalsky.winston.repository.fetch.YouTubeRequestRepository;
 import ca.metricalsky.winston.service.fetch.operation.FetchOperationHandler;
 import ca.metricalsky.winston.service.fetch.operation.FetchOperationHandlerFactory;
@@ -21,13 +19,11 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.http.HttpStatus;
-import org.springframework.test.util.ReflectionTestUtils;
 
 import java.time.OffsetDateTime;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.AdditionalAnswers.returnsFirstArg;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.verify;
@@ -45,33 +41,15 @@ class FetchServiceTest {
     @Mock
     private FetchOperationHandlerFactory fetchOperationHandlerFactory;
     @Mock
-    private FetchRequestMapper fetchRequestMapper;
-    @Mock
-    private FetchRequestRepository fetchRequestRepository;
-    @Mock
     private FetchRequestService fetchRequestService;
     @Mock
     private SsePublisher ssePublisher;
     @Mock
     private SsePublisherHolder ssePublisherHolder;
     @Mock
+    private YouTubeConfig youTubeConfig;
+    @Mock
     private YouTubeRequestRepository youTubeRequestRepository;
-
-    @Test
-    void save() {
-        var fetchRequest = new FetchRequest();
-        var fetchRequestEntity = buildFetchRequestEntity();
-
-        when(fetchRequestMapper.toFetchRequestEntity(fetchRequest))
-                .thenReturn(fetchRequestEntity);
-        when(fetchRequestRepository.save(fetchRequestEntity))
-                .thenAnswer(returnsFirstArg());
-
-        var fetchRequestId = fetchService.save(fetchRequest);
-
-        assertThat(fetchRequestId)
-                .isEqualTo(fetchRequestEntity.getId());
-    }
 
     @Test
     void fetchAsync() {
@@ -137,8 +115,8 @@ class FetchServiceTest {
 
     @Test
     void getAvailableQuota() {
-        ReflectionTestUtils.setField(fetchService, "dailyQuota", 10000);
-
+        when(youTubeConfig.getDailyRequestQuota())
+                .thenReturn(10000);
         when(youTubeRequestRepository.countAllByRequestedAtAfter(any(OffsetDateTime.class)))
                 .thenReturn(1);